refactor: 존재하지 않는 노선 삭제 시 예외처리 방식 변경

+++ src/main/java/wooteco/subway/line/application/LineService.java
@@ -91,10 +91,11 @@ public class LineService {
     }
 
     public void deleteLineById(Long id) {
-        int affectedRow = lineDao.deleteById(id);
-        if (affectedRow == 0) {
-            throw new NoSuchElementException("존재하지 않는 노선입니다.");
+        if (lineDao.isExistById(id)) {
+            lineDao.deleteById(id);
+            return;
         }
+        throw new NoSuchElementException("존재하지 않는 노선입니다.");
     }
 
     public void addLineStation(Long lineId, SectionRequest request) {
